requirements.txt
qrcode[pil]==7.4.2
reportlab==4.2.5
pillow==10.4.0
ttkbootstrap==1.20.1
